python/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data & ML
pandas>=2.1.4
numpy>=1.26.3
scipy>=1.12.0
scikit-learn>=1.4.0
torch>=2.1.2
yfinance>=0.2.36
pandas-ta>=0.3.14b0

# Vector database
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM clients
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.2

# Trading
alpaca-py>=0.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dateutil>=2.8.2

# Quant
quantlib>=1.32
